worker refactored
