+++ src/lambda/listener/index.ts
@@ -2,6 +2,7 @@ import { SNSClient } from "@aws-sdk/client-sns";
 import { S3Client } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { Octokit } from "@octokit/rest";
+
 import { createHandler } from "./handler";
 
 const pickEnv = <T extends string[]>(...keys: T): { [P in T[number]]: string } => {
